backend/requirements.txt
beautifulsoup4==4.14.2
fastapi==0.121.2
httpx==0.28.1
psycopg2_binary==2.9.11
python-dotenv==1.2.1
uvicorn==0.38.0
